Update T145_Jan-w49p20h4m2-10.krn
Deleted first bar as it is not in the Suter ID; deleted second time ending as this should have only been attached to the first bar, not the entire repeat section. Adjusted some beaming and fixed tremolando in the viola from 64th notes to 32nd notes, to better match Suter original.