Update T226_Sch-w35p124m568-569.krn
Adjusted natural harmonics to new encoding scheme in Humdrum.  Non-sounding notes are encoded as rests in a chord with the sounding note (fingered harmonics should always have three notes: the note pressed down, the note lightly touched to generate the harmonic (both of them encoded as rests at a certain pitch), and the sounding note.  Any note in that chord that is invisible should be given `yy` to hide it (not be omitted).   The middle note of the fingered harmonic chord will automatically be given a diamond-shaped notehead (so there are no layout parameters needed for this example anymore).

There is a problem with solid/open diamonds for half notes that should be fixed in verovio soon (but this is an 8th note, so no problem with the diamond shape).